move related enums to QueryGroup
